backend/foodgram/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
djoser==2.1.0
